Move class `CompactionNode` and `DifferSnapshotInfo` to separate files.
